requirements.txt
appdirs==1.4.4
babelfish==0.5.5
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dogpile.cache==0.9.2
enzyme==0.4.1
guessit==3.1.1
idna==2.9
pbr==5.4.5
pymediainfo==4.2.1
pynput==1.6.8
PyQt5==5.14.2
PyQt5-sip==12.7.2
pysrt==1.1.2
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
PyYAML==5.3.1
rarfile==3.1
rebulk==2.0.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
stevedore==1.32.0
subliminal==2.1.0
urllib3==1.25.9
